+++ src/com/intellij/idea/plugin/hybris/system/type/model/EnumType.java
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -37,6 +38,7 @@ public interface EnumType extends DomElement {
 
     String CODE = "code";
     String DYNAMIC = "dynamic";
+    String VALUE = "value";
 
     /**
      * Returns the value of the code child.
@@ -151,7 +153,7 @@ public interface EnumType extends DomElement {
      * @return the list of value children.
      */
     @NotNull
-    @SubTagList("value")
+    @SubTagList(VALUE)
     java.util.List<EnumValue> getValues();
 
     /**
@@ -159,7 +161,7 @@ public interface EnumType extends DomElement {
      *
      * @return created child
      */
-    @SubTagList("value")
+    @SubTagList(VALUE)
     EnumValue addValue();